api_serve/requirements.txt
alembic==1.0.0
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
Babel==2.6.0
billiard==3.3.0.23
celery==3.1.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.3.1
Flask==1.0.2
Flask-Admin==1.5.2
Flask-BabelEx==0.9.3
Flask-Cache==0.13.1
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
kombu==3.0.37
lxml==4.2.5
Mako==1.0.7
MarkupSafe==1.0
pycparser==2.19
PyJWT==1.6.4
PyMySQL==0.9.2
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
redis==2.10.6
requests==2.19.1
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.12
urllib3==1.23
vine==1.1.4
weixin-python==0.5.4
Werkzeug==0.14.1
WTForms==2.2.1
